Python code: Fix E201/E202 warnings

+++ gdal/swig/python/osgeo/gdal_array.py
@@ -174,7 +174,7 @@ import gdalconst
 import gdal
 gdal.AllRegister()
 
-codes = {   gdalconst.GDT_Byte      :   numpy.uint8,
+codes = {gdalconst.GDT_Byte      :   numpy.uint8,
             gdalconst.GDT_UInt16    :   numpy.uint16,
             gdalconst.GDT_Int16     :   numpy.int16,
             gdalconst.GDT_UInt32    :   numpy.uint32,
@@ -188,15 +188,15 @@ codes = {   gdalconst.GDT_Byte      :   numpy.uint8,
         }
 
 
-def OpenArray( array, prototype_ds = None ):
+def OpenArray(array, prototype_ds = None):
 
-    ds = OpenNumPyArray( array )
+    ds = OpenNumPyArray(array)
 
     if ds is not None and prototype_ds is not None:
         if type(prototype_ds).__name__ == 'str':
-            prototype_ds = gdal.Open( prototype_ds )
+            prototype_ds = gdal.Open(prototype_ds)
         if prototype_ds is not None:
-            CopyDatasetInfo( prototype_ds, ds )
+            CopyDatasetInfo(prototype_ds, ds)
 
     return ds
 
@@ -231,32 +231,32 @@ def GDALTypeCodeToNumericTypeCode(gdal_code):
     return flip_code(gdal_code)
 
 
-def LoadFile( filename, xoff=0, yoff=0, xsize=None, ysize=None,
+def LoadFile(filename, xoff=0, yoff=0, xsize=None, ysize=None,
               buf_xsize=None, buf_ysize=None, buf_type=None,
               resample_alg = gdal.GRIORA_NearestNeighbour,
-              callback=None, callback_data=None ):
-    ds = gdal.Open( filename )
+              callback=None, callback_data=None):
+    ds = gdal.Open(filename)
     if ds is None:
         raise ValueError("Can't open "+filename+"\n\n"+gdal.GetLastErrorMsg())
 
-    return DatasetReadAsArray( ds, xoff, yoff, xsize, ysize,
+    return DatasetReadAsArray(ds, xoff, yoff, xsize, ysize,
                                buf_xsize=buf_xsize, buf_ysize=buf_ysize, buf_type=buf_type,
                                resample_alg=resample_alg,
-                               callback = callback, callback_data = callback_data )
+                               callback = callback, callback_data = callback_data)
 
 
-def SaveArray( src_array, filename, format = "GTiff", prototype = None ):
-    driver = gdal.GetDriverByName( format )
+def SaveArray(src_array, filename, format = "GTiff", prototype = None):
+    driver = gdal.GetDriverByName(format)
     if driver is None:
         raise ValueError("Can't find driver "+format)
 
-    return driver.CreateCopy( filename, OpenArray(src_array,prototype) )
+    return driver.CreateCopy(filename, OpenArray(src_array,prototype))
 
 
-def DatasetReadAsArray( ds, xoff=0, yoff=0, win_xsize=None, win_ysize=None, buf_obj=None,
+def DatasetReadAsArray(ds, xoff=0, yoff=0, win_xsize=None, win_ysize=None, buf_obj=None,
                         buf_xsize = None, buf_ysize = None, buf_type = None,
                         resample_alg = gdal.GRIORA_NearestNeighbour,
-                        callback=None, callback_data=None ):
+                        callback=None, callback_data=None):
     """Pure python implementation of reading a chunk of a GDAL file
     into a numpy array.  Used by the gdal.Dataset.ReadAsArray method."""
 
@@ -269,12 +269,12 @@ def DatasetReadAsArray( ds, xoff=0, yoff=0, win_xsize=None, win_ysize=None, buf_
         return None
 
     if ds.RasterCount == 1:
-        return BandReadAsArray( ds.GetRasterBand(1), xoff, yoff, win_xsize, win_ysize,
+        return BandReadAsArray(ds.GetRasterBand(1), xoff, yoff, win_xsize, win_ysize,
                                 buf_xsize = buf_xsize, buf_ysize = buf_ysize, buf_type = buf_type,
                                 buf_obj = buf_obj,
                                 resample_alg = resample_alg,
                                 callback = callback,
-                                callback_data = callback_data )
+                                callback_data = callback_data)
 
     if buf_obj is None:
         if buf_xsize is None:
@@ -287,7 +287,7 @@ def DatasetReadAsArray( ds, xoff=0, yoff=0, win_xsize=None, win_ysize=None, buf_
                 if buf_type != ds.GetRasterBand(band_index).DataType:
                     buf_type = gdalconst.GDT_Float32
 
-        typecode = GDALTypeCodeToNumericTypeCode( buf_type )
+        typecode = GDALTypeCodeToNumericTypeCode(buf_type)
         if typecode == None:
             buf_type = gdalconst.GDT_Float32
             typecode = numpy.float32
@@ -308,21 +308,21 @@ def DatasetReadAsArray( ds, xoff=0, yoff=0, win_xsize=None, win_ysize=None, buf_
         if buf_obj.shape[0] != ds.RasterCount:
             raise ValueError('Array should have space for %d bands' % ds.RasterCount)
 
-        datatype = NumericTypeCodeToGDALTypeCode( buf_obj.dtype.type )
+        datatype = NumericTypeCodeToGDALTypeCode(buf_obj.dtype.type)
         if not datatype:
             raise ValueError("array does not have corresponding GDAL data type")
         if buf_type is not None and buf_type != datatype:
             raise ValueError("Specified buf_type not consistent with array type")
         buf_type = datatype
 
-    if DatasetIONumPy( ds, 0, xoff, yoff, win_xsize, win_ysize,
-                       buf_obj, buf_type, resample_alg, callback, callback_data ) != 0:
+    if DatasetIONumPy(ds, 0, xoff, yoff, win_xsize, win_ysize,
+                       buf_obj, buf_type, resample_alg, callback, callback_data) != 0:
         return None
 
     return buf_obj
 
 
-def BandReadAsArray( band, xoff = 0, yoff = 0, win_xsize = None, win_ysize = None,
+def BandReadAsArray(band, xoff = 0, yoff = 0, win_xsize = None, win_ysize = None,
                      buf_xsize=None, buf_ysize=None, buf_type=None, buf_obj=None,
                      resample_alg = gdal.GRIORA_NearestNeighbour,
                      callback=None, callback_data=None):
@@ -342,12 +342,12 @@ def BandReadAsArray( band, xoff = 0, yoff = 0, win_xsize = None, win_ysize = Non
         if buf_type is None:
             buf_type = band.DataType
 
-        typecode = GDALTypeCodeToNumericTypeCode( buf_type )
+        typecode = GDALTypeCodeToNumericTypeCode(buf_type)
         if typecode == None:
             buf_type = gdalconst.GDT_Float32
             typecode = numpy.float32
         else:
-            buf_type = NumericTypeCodeToGDALTypeCode( typecode )
+            buf_type = NumericTypeCodeToGDALTypeCode(typecode)
 
         if buf_type == gdalconst.GDT_Byte and band.GetMetadataItem('PIXELTYPE', 'IMAGE_STRUCTURE') == 'SIGNEDBYTE':
             typecode = numpy.int8
@@ -365,23 +365,23 @@ def BandReadAsArray( band, xoff = 0, yoff = 0, win_xsize = None, win_ysize = Non
         if buf_ysize is not None and buf_ysize != shape_buf_ysize:
             raise ValueError('Specified buf_ysize not consistent with array shape')
 
-        datatype = NumericTypeCodeToGDALTypeCode( buf_obj.dtype.type )
+        datatype = NumericTypeCodeToGDALTypeCode(buf_obj.dtype.type)
         if not datatype:
             raise ValueError("array does not have corresponding GDAL data type")
         if buf_type is not None and buf_type != datatype:
             raise ValueError("Specified buf_type not consistent with array type")
         buf_type = datatype
 
-    if BandRasterIONumPy( band, 0, xoff, yoff, win_xsize, win_ysize,
-                          buf_obj, buf_type, resample_alg, callback, callback_data ) != 0:
+    if BandRasterIONumPy(band, 0, xoff, yoff, win_xsize, win_ysize,
+                          buf_obj, buf_type, resample_alg, callback, callback_data) != 0:
         return None
 
     return buf_obj
 
 
-def BandWriteArray( band, array, xoff=0, yoff=0,
+def BandWriteArray(band, array, xoff=0, yoff=0,
                     resample_alg = gdal.GRIORA_NearestNeighbour,
-                    callback=None, callback_data=None ):
+                    callback=None, callback_data=None):
     """Pure python implementation of writing a chunk of a GDAL file
     from a numpy array.  Used by the gdal.Band.WriteArray method."""
 
@@ -394,20 +394,20 @@ def BandWriteArray( band, array, xoff=0, yoff=0,
     if xsize + xoff > band.XSize or ysize + yoff > band.YSize:
         raise ValueError("array larger than output file, or offset off edge")
 
-    datatype = NumericTypeCodeToGDALTypeCode( array.dtype.type )
+    datatype = NumericTypeCodeToGDALTypeCode(array.dtype.type)
 
 # if we receive some odd type, like int64, try casting to a very
 # generic type we do support (#2285)
     if not datatype:
-        gdal.Debug( 'gdal_array', 'force array to float64' )
-        array = array.astype( numpy.float64 )
-        datatype = NumericTypeCodeToGDALTypeCode( array.dtype.type )
+        gdal.Debug('gdal_array', 'force array to float64')
+        array = array.astype(numpy.float64)
+        datatype = NumericTypeCodeToGDALTypeCode(array.dtype.type)
 
     if not datatype:
         raise ValueError("array does not have corresponding GDAL data type")
 
-    return BandRasterIONumPy( band, 1, xoff, yoff, xsize, ysize,
-                                array, datatype, resample_alg, callback, callback_data )
+    return BandRasterIONumPy(band, 1, xoff, yoff, xsize, ysize,
+                                array, datatype, resample_alg, callback, callback_data)
 
 
 def RATWriteArray(rat, array, field, start=0):
@@ -457,7 +457,7 @@ def RATReadArray(rat, field, start=0, length=None):
     return RATValuesIONumPyRead(rat, field, start, length)
 
 
-def CopyDatasetInfo( src, dst, xoff=0, yoff=0 ):
+def CopyDatasetInfo(src, dst, xoff=0, yoff=0):
     """
     Copy georeferencing information and metadata from one dataset to another.
     src: input dataset
@@ -469,27 +469,27 @@ def CopyDatasetInfo( src, dst, xoff=0, yoff=0 ):
 
     """
 
-    dst.SetMetadata( src.GetMetadata() )
+    dst.SetMetadata(src.GetMetadata())
 
 
 #Check for geo transform
     gt = src.GetGeoTransform()
     if gt != (0,1,0,0,0,1):
-        dst.SetProjection( src.GetProjectionRef() )
+        dst.SetProjection(src.GetProjectionRef())
 
         if (xoff == 0) and (yoff == 0):
-            dst.SetGeoTransform( gt  )
+            dst.SetGeoTransform(gt)
         else:
             ngt = [gt[0],gt[1],gt[2],gt[3],gt[4],gt[5]]
             ngt[0] = gt[0] + xoff*gt[1] + yoff*gt[2];
             ngt[3] = gt[3] + xoff*gt[4] + yoff*gt[5];
-            dst.SetGeoTransform( ( ngt[0], ngt[1], ngt[2], ngt[3], ngt[4], ngt[5] ) )
+            dst.SetGeoTransform((ngt[0], ngt[1], ngt[2], ngt[3], ngt[4], ngt[5]))
 
 #Check for GCPs
     elif src.GetGCPCount() > 0:
 
         if (xoff == 0) and (yoff == 0):
-            dst.SetGCPs( src.GetGCPs(), src.GetGCPProjection() )
+            dst.SetGCPs(src.GetGCPs(), src.GetGCPProjection())
         else:
             gcps = src.GetGCPs()
 #Shift gcps
@@ -506,7 +506,7 @@ def CopyDatasetInfo( src, dst, xoff=0, yoff=0 ):
                 new_gcps.append(ngcp)
 
             try:
-                dst.SetGCPs( new_gcps , src.GetGCPProjection() )
+                dst.SetGCPs(new_gcps , src.GetGCPProjection())
             except:
                 print("Failed to set GCPs")
                 return
